Implement better mouse support.
This initial version is very messy and the timing for scroll events needs further changes
as a simple trigger and maximum yielded more desirable results than what is here right now
using a timeout.

Click events are now generated on mouse up. Use of a delay to check for double-click will come
later if at all. Proper Application Events need to be realized, but this is a start.

Drag and drop support will likely be non-traditional as other features handle the use case for moving
text around.

Additional changes:
	- Indicator rendering now identifies inverted start and stop.
	However, it does not swap the colors of the cursor as it should.
	- Bisect position indicator.
	- Various display bugs involving the lack of signalling that change occurred.
	- Preliminary support for adjustable pane counts.
	- Color changes to leverage 24-bit colors.