Use event based source and destination for terminal I/O.
On macOS, opening `/dev/tty` directly gives a file descriptor that
behaves as a regular data file: the file descriptor is always readable
according to kevent. However, accessing the actual terminal device
associated with the process does not exhibit that behavior. This can
be done with either an explicit parameter or by analyzing stdio using
&unix.ttyname.

Rewrite the terminal flow to use event I/O while leaving the thread
based path available in case it proves problematic. Event I/O is
primarily intended for large transfers, and terminal input is often
a series of few-byte transfers.