Isolate event handlers from application classes.
The initial step of a larger scale change. The ultimate goal being
the refactoring of the event handlers to work with specific Interfaces
where the action's endpoint is selected using a combination of the
application's state and the configured binding that invoked the action.

Originally, event handlers were implemented inline for the usual
inheritance benefits. However, the infrequent inheritance use, line count,
and the ambiguity of purpose has made the leveraged value irrelevant. While
crude, the given Index implementation can perform method substitution with
locally bound indexes(currently not used). In the future, if event handling
overrides are desired(external to/before the application), it will still be
possible.