Begin isolating the Resource element from Refraction.
Conceptually, this was being done, but begin making it explicit by
defining a Resource class that holds the log, lines, and reference
for a file. Move the system context into reference as well so that
a single point can be used in order to perform resource I/O.

Refraction will likely integrate View fully at some point and the
interfaces to the elements should become more sensible with a
clearer purpose. Sessions attach Frames, Frames attach Refractions,
Refractions attach Resources.