Isolate Execution context from System identities and refactor session resource methods.
Formalize the execution environment that will dispatch system operations as
an element (of a session). The systems recognized by a session are keyed by
the identities which are held by References. When resource operations are performed,
the system will be resolved using the session's table avoiding the potential for
reference cycles.

The minor refactoring of session resource methods avoids unnecessary resolution
of Resource instances and attempts to make the method names more consistent and clear
with respect the intention of process local resources.