Add configuration project to house the settings for the syntax editor.
The plan is to allow different means for designating the configuration,
but with (versioned) Application Instructions being the primary means.

The current implementation borrows from syntax.ia in controls.py, but
it is intended to keep the configuration data dependency free with
exception to __init__ where some loading (entry point) interfaces are
implemented.