Implement Reformulations and use the configuration project.
Reformulations serve as a sort of rendering context for the entire
I/O stack of resources from the encoded binary all the way to the
cells dispatched for display into Screens. Much of this existed in
the syntax.format module, but loosely coupled on Refractions. Now,
resource types are defined as Reformulations which are, in turn,
inherited by the Resource and then the Refraction. Controls over
display will be possible at every level.

Session now takes a configuration as its primary, first, parameter.
Currently, this is just the terminal.configuration package module, but
a typed form, that terminal.cfg conforms to, is likely.

syntax.keyboard has been moved into configuration.controls, which will likely
be moved into a qwerty specific module that will be dumped in cfg.controls.

Much of reformulations is transitional as Refraction.structure and
Refraction.render are still being used.