Integrate (resource) element modification functions on elements.Resource.
The "sparse" method approach offers some advantage, but Resource is the
canonical container and easily constructed so tests will not suffer. However,
these methods may be relocated once again if isolation is needed by the
future cursor element.

While here, try to establish lo and co as primary identifiers for describing
the line-offset and codepoint-offset. Back to avoiding `ln` as types.Line
instances will prefer that identifier.

Some regressions may be present here with differential rendering, but
ignored as a delta.Log based rewrite is expected.