Constrain the cursor to the indexable lines or zero.
Prior, the cursor could be positioned offscreen and needed to
in order to place the vertical stop (exclusive) at the end of the file.

To compensate, change (place/stop) to be inclusive with the cursor's line
so that the all lines may be selected by the range. Visually, the stop indicators
are still placed on the limit and will remain so that single line ranges may
be represented with the least conflict.