Implement project local resource structured line storage.
Internally, lines were stored, nearly, as the raw strings that were read.
The advantage of this format is, predominantely, that offset alignment
between the raw line, the structured fields, and the phrase is
naturally achieved. However, compensation is no longer difficult. Originally,
a structured line was used for local memory storage, but proved inconvenient
at the time due to issues with alignment. The, nearly, raw string replaced
a structured line(indentation level and line content) to help with phrase rendering.

Now, raw strings are still being stored in elements, but are, often, interpreted
and written through &.types.Line instances. These raw strings leverage an internal
format to make sequencing and structuring (to/from &.types.Line) as trivial as
possible while still allowing &.delta manipulations to be used for *all* changes.
No structure aware difference algorithm is needed.

Once again, indentation is, essentially, independent of line content. However,
unlike before, it is stored as a single character at the beginning of the formatted
line stored in elements.

This formatting intends to provide for line extensions that allow for annotations,
line local formatting, and type overrides to be applied.