Correct or reduce cursor blinking and push source deltas on commit.
With parallel views held by the session/frame, Resource changes were,
essentially, polled by keeping track of the commit on the Refraction.
Polling itself, list slice, was inexpensive, but differential rendering
attempted to handle arbitrary (source) delta sequences complicating
its implementation. Not to mention responding to scrolls at the end.

Differential rendering now works with a single source delta per-call
and is invoked for each view connected, via weak reference, to the
Resource.

Most display deltas are now appended to Frame.deltas which is connected
to Refractions on Refraction.configure. Currently, this is flushed at
the end of an Application Instruction with the cursor indicator.

Frame.indicate has been refactored into a few methods allowing Refractions
to define how their cursor is displayed. Cursor blinking has been reduced
or eliminated entirely by using the same technique as before but relying
on the (Frame) deferred display deltas. By doing so earlier in the pipeline,
any blinking is eliminated as the invalidations are performed with the
final cell image. A redundant pixel update is, currently, performed, but it
renders the same pixels so no change should be perceived.

Potentially, this can shifted back downstream, macOS, with the employment
of double pixel buffers to control framing(as many IOSurface users do).

- Minor refactoring of delta.Log.undo and delta.Log.redo to compensate
for how source deltas are applied by Resource. Returns deltas instead
of applying them.
- Introduce types.Status for some polling based updates. (maybe short lived)
- Change Refraction.scroll to enqueue display deltas rather than relying
on DR.
- Prefix methods adjusting the View Image with (vi_). Likely, this is
going back into its own object (once again), but in &.cells.text.
- Eliminate Frame.parallels in favor of weak references in Resource views.
- Frame.deltas now stores display deltas rather than Refractions to be polled.
- Add trace command to toggle event tracing at runtime.

I'm not fond of this use of weak references on Resources, but managing source
deltas with generators alongside of display deltas would require type switching
in the receiver or forcing Application Instructions to be performed relative
to a View with parallel Refractions assigned to the View? Supporting parallel
views complicates the return dramatically.

While the back references currently muddy the instruction flow, it is an
accurate representation of how changes, currently, need to be propagated.