Isolate the view image from the Refraction, once again.
Originally called "View", this structure was intended to be an
application independant buffer for division displays. With the
expectation of Reflections using Screen instances, this became
an unlikely invariant.

However, the coupling of a Phrase sequence and cell offset positions
is very useful regardless of an application's structure. So, place an
implementation alongside of the Phrase as a common tool for managing
a view consisting of a Phrase sequence.

There is still some possible work to be done here as it may be
appropriate for the Image to play a larger role in differential
rendering.