Implement logged cursor displacement in favor of splice compensation.
Currently, line changes have no influence on the horizontal cursor. However,
some influence is needed for splice_text to function as it might be expected
to. Cursor displacment records allows commits to move the cursor independently
of the physical changes, but relative to the context of the transaction.

Alternatively, delta.Lines may integrate this in the future or order to describe
the phsyical change more accurately.

Additionally, checkpoint on the log directly in Insertion.final in order to avoid
a logging bug. When committing (src.checkpoint) unconditionally, insertions were,
sometimes, being performed redundantly when splice_text was used. The exact
origin of this is unknown, but is likely related to combining records.