Integrate "cross" methods into the element that provides the necessary scope.
Prior, the application instructions were implemented as a set of sparse methods taking
the session, frame, and refraction as the primary parameters. The scope is far
too wide for most instructions with the majority being cursor methods.

While this offers confusion with cursor methods sometimes being present on the
Session, it is trivial to recognize the appropriate location for a method by
its scope requirements.

- Eliminates redirections in favor of a delta modifier injected early in event handling.
. Allows for normal bindings to be used to select a target (y) distribution method.
- Eliminates intercepts in favor of direct access to application instructions.