Implement double buffering with whole frame memory copies.
This may be inefficient for the normal case as little is often changing
on the screen. However, implementing double buffering with delta replays was
a more significant change, so keep it simple for the moment.

The recently reorganized cursor rendering code eliminated blinking with the
exception of scrolling where the pixel replication operation forced the
cursor restoration cells to be briefly displayed while the pixels were being
copied. This allows the displayed frame to remain static while the other is
being updated for display.