Implement status modifiers for controlling Return key behavior.
The goal of this solution is to avoid switching on modifiers inside
selected methods and to avoid layers of (keyboard) modes.

The cost of this solution is a full combination of Return bindings for the
session/activate instruction so that they can be properly forwarded.

session/cancel is now handled this way as well for prompt retention and,
likely, location hiding.

Additionally:

- Actually break the line on Return when in insert mode.
- Add a skeleton transcript type which prefers retained prompts.
- Enable command substitution in all division sectors(prompt and location).