Stop filling layout in redistribute and include borders in n-allocation verticals.
Include the space of the missing borders to allow visual alignment with frames
that have n-single unit verticals.

The removed layout extension wasn't working as it was originally. Units being
a sum of the division count, it was not even adding the right unit-width to
identify that expansion should be performed. At this point, it's probably better
to relocate that fill mechanism elsewhere and leave redistribute to focus on
excess inheritance and the calculation of the vertical positions.