Refactor application instructions to take flexible arguments.
This change allows AI implementations to reside on the most relevant class,
and nearly brings them back to being crossmethods. However, the selectors
provide a much needed specialization so that implementations aren't just
taking the entire session scope. This will be important for testing in the
future where mocking a Session or Frame could be cumbersome and fragile.

Some methods have been moved, but most have been kept in place in order
to keep the initial change to merely compensate for the signature differences.