Relocate file I/O operations to Execution and isolate process local resources into Directory.
Directory attempts to, further, formalize access to process local resources.
Prior, it was just a dictionary on session with some higher level methods
directly integrated in Session. The Session.resources dictionary
lasted as long as it did as the provided abstraction has little utility.
However, fragmenting the session scope is desirable to limit cyclic access
to the, essentially, process global data.

The store and load methods relocated to system.Execution are still hardcoded
to perform host I/O. Future implementations will use system relative subprocesses
to read and write sources. Much of this commit is merely preparation for this.