Prefer 'seek' for some consistency with cursor operations.
It may not be appropriate for complete consistency as panning
is strictly cell oriented as opposed to the cursor's character unit.

The similarity is mostly intended to aid in the recollection of shorthands.
That is, `csln` and `vsln` have one differing character for similar operations.

`view/scroll` and `view/pan` are maintained for the moment due to
the, current, limitations of direct instruction dispatch in the cocoa device.