Add support for literal space characters.
Allocate a character from a Unicode PUA to represent a literal
space on the prompt. Eventually, this will be used in conjunction
with quoted literals where the escaped literal space will not be
translated in order to allow the PUA character to be referenced
without an additional escape character.

In part, this is a short term solution given the absence of a parser
for the prompt. However, it is intended to be a long term feature
for representing literal spaces outside of quotations.