Implement procedures, compositions, and redirects.
- Work type implements/replaces Refraction.system_execution_status.
- system.Completion no longer refers to a Refraction.
- Event link context now holds the Work (weak) reference and cursor index.
- Work.interrupt currently presumes host as the system is only accessible
via the generator state.
- Work instances are only held by ExecutionStatus annotations.
- Ignore prompt close when performing a sole `cd`.

The implementation is short of a proper abstraction. Notably, procedures
have to be executed by a virtual process (Work.cursors). Seamless upon
pg_execute, but with some difficulty when spawned from a composition where
the Work's procedure set needs to be expanded to handle the sequence.
These virtual or real processes are necessary and may be improved if they
could be launched with greater transparency than they currently are.