Prompt for resource saves.
Prior to redirects, the prompt had no way of referencing the
PLR. So, location actuation was overloaded to handle save as well
as location switches. This overload was managed by switching on
the directory annotation's title. A working stopgap, but just
that as the mechanism is poorly defined and easily broken by
switching the annotation.

With redirects, tee can now be given the entire file without
a special case and saving a file is just another issued command.
This comes with a single drawback of having to fully buffer the
lines prior to the execution of the save process (tee) as misdirects
need to support safe replacements. However, this also protects against
out-of-scope changes being inadvertantly written.