Implement Resource revisions for prompt and location history.
Locations and Prompts are, currently, conceptual from a class perspective; regular Refractions
with a custom syntax type provide the desired features for locating resources and executing
commands. This may not always be the case in the future, but, for now, it is a priority
to maintain a consistent foundation where possible.

Revisions are the general mechanism used to implement location and prompt histories.

While intended for prompts and locations, process instructions may be used to manage revisions
for content as well. Major points in history may be explicitly kept for later reference by the
user. As maintenance procedures become implicit, this will be a necessity as default limits
will be applied to deltas and far histories can be lost. Content revisions allow users
to hold on to specific versions for arbitrary durations within the scope of the session.

- Add browser familiar key binds for navigating location histories and prompt histories.
- Allow prompt history to be scrolled while the content is focused.
- Add prompts.history_limit for trimming history during maintenance.
- Scan and reduce revision deltas during maintenance.
- Propogate the process' system identity to Prompting for easier access.
- Split pg_open. Require callers to use pg_configure_command directly.
- Correct footer resize bug by updating the frame's copy of the areas.
- Add pg_process interface for simplifying process command setup. System commands cannot
be handled so easily as they depend on the resource's system identity.
- Seek the end when attaching reporting refractions. Currently, transcripts are cursor
sensitive, so make it slightly more ergonomic by preferring the end of the file.