Replace return views with view stacks.
Return views were intentionally limited, and were going to be awkward to use without
additional efforts. So discard the idea entirely in favor of stacks and location histories.

- Add view.Structure type for holding a division's elements(location, content, prompt).
- Add reference selector for division (view index).
- Adjust `Resource > Close` to perform View closing instead of resource closing as Views
are a closer analog to Windows, the usual meaning of Command-W.
- Clarify Frame.switch as Frame.switch_division.
- Protect against some errant display instructions with frame_visible checks.

Commit is intentionally missing needed adjustments to retention and merely ignores
and empties returnview values in the interim.