Rename "io" project to "device".
The device is the project's only purpose. The plan was to conflate the devices so that only one
dependency needed to be referenced to simplify application linkage, but addressing individual
implementations will likely be necessary as multiple device supporting platforms may be usable
within a single system. Notably, web and xterm devices are likely to be supported on all systems.