Improve rendering performance by patching on image dispatch.
Full copies of the pixel image works on modern systems, but significantly hurt the typing
experience on older machines. Adjust the invalidation strategy to include a level for
image dispatch to use for tracking which updates have been dispatched to the display.

However, when over a third of the pixel image has been updated, continue copying full.