Implement the mirror terminal device and Reflection view.
While projections are intended to be a primary means for working with remote systems, it
is not always preferred as the desired session may exist on the remote. Additionally, Reflections
provide the necessary support for interfacing with isolated terminal applications written in
any language. Notably, legacy terminal emulation.

- Reflection view for representing the client process.
- Host binary Controller and Display protocols. (trivial raw struct transfers)
- Common CellMatrix element for Refraction and Reflection.
- Correct and improve resize operations.
- Adjust device implementations to use the application's provided screen allocation.
- Add device control keys as an eventual replacement for most application instruction keys.
Nearly just a rename, but will eliminate application specific instructions for generic handles.
- Correct Area intersections to be zero volume or only selecting areas within the bounds.
- Zero (Area) dimensions are now on the destination of replication requests.
Prior, the source held the zeros but mirror's protocol needs the zeros on the leading Area.