Integrate display synchronization into replicate_cells for macos-cocoa.
With some devices not requiring synchronization after replicate, performing the operation
in the application is no longer desired. Additionally, use of synchronize was missing some
concurrency opportunities by waiting for the entire replication operation to complete.

By only blocking on updatePixels, display deltas may continue to be processed while
pixel replication is being performed as it only depends on the pixel image.